+++ packages/@lou.codes/parsers/src/undefineNaN.ts
@@ -20,5 +20,5 @@ import type { Maybe } from "@lou.codes/types";
 export const undefineNaN = <Number extends number>(
 	// eslint-disable-next-line @typescript-eslint/ban-types
 	number: Number,
-	// eslint-disable-next-line @typescript-eslint/ban-types
+	// eslint-disable-next-line @typescript-eslint/ban-types, @typescript-eslint/no-unnecessary-type-assertion
 ) => (Number.isNaN(number) ? undefined : number) as Maybe<Number>;