+++ packages/@lou.codes/utils/src/constructDate.ts
@@ -1,4 +1,4 @@
-import type { ReadOnly } from "@lou.codes/types";
+import type { ReadOnly, Unbound } from "@lou.codes/types";
 import { construct } from "./construct.js";
 
 /**
@@ -8,4 +8,4 @@ import { construct } from "./construct.js";
  */
 export const constructDate = construct(Date) as (
 	date?: ReadOnly<Date> | number | string,
-) => Date;
+) => Unbound<Date>;