+++ packages/@lou.codes/utils/src/constructProxy.ts
@@ -1,3 +1,4 @@
+import type { Unbound } from "@lou.codes/types";
 import { construct } from "./construct.js";
 
 /**
@@ -8,4 +9,4 @@ import { construct } from "./construct.js";
 export const constructProxy = construct(Proxy) as <Target extends object>(
 	target: Target,
 	handler: ProxyHandler<Target>,
-) => Target;
+) => Unbound<Target>;