+++ packages/@lou.codes/utils/src/constructSet.ts
@@ -1,3 +1,4 @@
+import type { Unbound } from "@lou.codes/types";
 import { construct } from "./construct.js";
 
 /**
@@ -8,4 +9,4 @@ import { construct } from "./construct.js";
 export const constructSet = construct(Set) as <Value>(
 	iterable?: Iterable<Value>,
 	// eslint-disable-next-line functional/prefer-readonly-type
-) => Set<Value>;
+) => Unbound<Set<Value>>;