+++ packages/@lou.codes/utils/src/constructWeakSet.ts
@@ -1,3 +1,4 @@
+import type { Unbound } from "@lou.codes/types";
 import { construct } from "./construct.js";
 
 /**
@@ -8,4 +9,4 @@ import { construct } from "./construct.js";
 export const constructWeakSet = construct(WeakSet) as <Value extends object>(
 	iterable?: Iterable<Value>,
 	// eslint-disable-next-line functional/prefer-readonly-type
-) => WeakSet<Value>;
+) => Unbound<WeakSet<Value>>;