+++ scripts/formatPairedDocs.js
@@ -5,25 +5,49 @@ import { getGroupedFiles } from "./getGroupedFiles.js";
 export const formatPairedDocs = () =>
 	getGroupedFiles().then(filePairs =>
 		filePairs.map(
-			([readmeFile, packageFile, typeDocFile]) =>
-				/** @type {const} */ ([
-					typeDocFile.path.replace(
-						TYPE_DOC_FILES_DIRECTORY,
-						ASTRO_TARGET,
-					),
+			([
+				readmeFile,
+				packageFile,
+				[typeDocFile, ...moduleTypeDocFiles],
+			]) => {
+				const { groups: { moduleName = "", packageName = "" } = {} } =
+					typeDocFile.path.match(
+						/\/(?<packageName>[^/\\]+)\.(?<moduleName>[^/\\]+)\.md/u,
+					) ?? {};
+				const isModule = moduleName !== "";
+
+				return /** @type {const} */ ([
+					typeDocFile.path
+						.replace(TYPE_DOC_FILES_DIRECTORY, ASTRO_TARGET)
+						.replace(
+							`/${packageName}.${moduleName}.md`,
+							`/${packageName}/${moduleName}.md`,
+						)
+						.toLocaleLowerCase(),
 					`${frontMatter({
 						description: packageFile.description,
-						title: packageFile.name,
+						title:
+							isModule ?
+								`${packageFile.name} (${moduleName})`
+							:	packageFile.name,
 					})}\n\n${
-						readmeFile.content
-					}\n\n<!-- Start of auto-generated code by TypeDoc -->\n\n${typeDocFile.content
+						isModule ? "" : readmeFile.content
+					}\n\n<!-- Start of auto-generated code by TypeDoc -->\n\n${[
+						typeDocFile.content,
+						...moduleTypeDocFiles.map(
+							moduleTypeDocFile => moduleTypeDocFile.content,
+						),
+					]
+						.join("\n")
 						.split("\n")
 						.slice(2)
 						.join("\n")
+						// Make all references to the TypeDoc file relative
 						.replaceAll(
 							typeDocFile.path.split("/").at(-1) ?? "",
 							"",
 						)
+						// Make the "defined in" nicer.
 						.replaceAll(
 							/#### Defined in\n\n\[.+\]\((?<path>.+)\)/gu,
 							"> [View source]($1)",
@@ -38,16 +62,27 @@ export const formatPairedDocs = () =>
 							/#### Defined in\npackages\/(?<sourcePackageName>@lou\.codes\/.+)\/node_modules\/(?<packageName>@lou\.codes\/.+)\/dist\/(?<filename>.+)\.d\.ts:\d+/gu,
 							"> [View source](https://github.com/loucyx/lou.codes/blob/main/packages/$1/src/$2.ts)",
 						)
+						// Make absolute references to the site relative
 						.replaceAll("https://lou.codes/", "/")
 						.replaceAll("https://lou.codes", "/")
+						// Turn all question marks next to properties into " (optional)"
 						.replaceAll(
 							/(?<optionalPropertyEnd>\?`)(?<spaces> +\|)/gu,
 							" (optional)`$2",
 						)
+						// Make function headers nicer
 						.replaceAll(
 							/(?<symbol>▸|Ƭ) (?<code>[^\n]*(?:\n[^\n]+)*)\n\n/gu,
 							`<div class="font-mono text-sm">\n\n$1 $2\n\n</div>\n\n`,
-						),
-				]),
+						)
+						// Remove modules links because we are inlining them
+						.replaceAll(
+							/## Modules\n\n-(?<moduleLink>.+\n)+\n/gu,
+							"",
+						)
+						// Remove module title/description
+						.replaceAll(/(?<moduleTitle># Module:.+\n)\n.+/gu, ""),
+				]);
+			},
 		),
 	);