+++ packages/@lou.codes/diff/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## 2.0.1
+
+### Patch Changes
+
+-   ⬆️ upgrade `@lou.codes/types` to `v1.2.0`
+
 ## 2.0.0
 
 ### Major Changes