+++ packages/@lou.codes/predicates/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## 1.0.14
+
+### Patch Changes
+
+-   ⬆️ upgrade `@lou.codes/types` to `v1.2.0`
+-   ⬆️ upgrade `@lou.codes/parsers` to `v1.0.14`
+
 ## 1.0.13
 
 ### Patch Changes