+++ packages/@lou.codes/prompts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## 1.0.15
+
+### Patch Changes
+
+-   ⬆️ upgrade `@lou.codes/types` to `v1.2.0`
+
 ## 1.0.14
 
 ### Patch Changes